Moving tests to examples since multichip pytorch tester is not implemented
